refactor: move Rotation into one separate module and add test for it
