+++ Cargo.toml
@@ -56,6 +56,7 @@ serde_json = { version = "1.0", optional = true }
 time = { version = "0.3", features = [
   "formatting",
   "parsing",
+  "macros",
 ], optional = true }
 
 ## Fastrace dependencies
