+++ src/append/rolling_file/mod.rs
@@ -18,11 +18,12 @@ pub use non_blocking::NonBlockingBuilder;
 pub use non_blocking::WorkerGuard;
 pub use rolling::RollingFileWriter;
 pub use rolling::RollingFileWriterBuilder;
-pub use rolling::Rotation;
+pub use rotation::Rotation;
 
 mod append;
 mod non_blocking;
 mod rolling;
+mod rotation;
 mod worker;
 
 #[derive(Debug)]
